+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## not released yet
 
+#### Improvements
+- disable aws sdk logger if log level is not "trace"
 
 ## v2.0.0 - 4 Jul 2022